Extract FlamechartDetailView to its own file
